refactor: moved utils for pagination into seperate module -Created new module 'pagination.py' and moved utils to it. -This makes code more readable -Also deleted some unused and old files that are no longer used. Tell me what you think...
